requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
psycopg2-binary  # solo necesario si usas PostgreSQL
pydantic[email]
pytz